refactor(no-element): get rid of useNoElementRender composable and NoElement component in favour of NoAnimation component
