+++ packages/x-components/src/components/modals/base-modal.vue
@@ -31,14 +31,11 @@
 </template>
 
 <script lang="ts">
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  import Vue, { defineComponent, nextTick, onBeforeUnmount, onMounted, ref, watch } from 'vue';
-  import { useDebounce } from '../../composables/use-debounce';
+  import { defineComponent, nextTick, onBeforeUnmount, onMounted, ref, watch } from 'vue';
+  import { useDebounce } from '../../composables';
   import { AnimationProp } from '../../types';
-  import { getTargetElement } from '../../utils/html';
-  import Fade from '../animations/fade.vue';
-  import { NoElement } from '../no-element';
-  import { FOCUSABLE_SELECTORS } from '../../utils/focus';
+  import { getTargetElement, FOCUSABLE_SELECTORS } from '../../utils';
+  import { Fade, NoAnimation } from '../animations';
 
   /**
    * Base component with no XPlugin dependencies that serves as a utility for constructing more
@@ -48,7 +45,6 @@
    */
   export default defineComponent({
     name: 'BaseModal',
-    components: { NoElement },
     props: {
       /** Determines if the modal is open or not. */
       open: {
@@ -72,7 +68,7 @@
       /** Animation to use for opening/closing the modal.This animation only affects the content. */
       animation: {
         type: AnimationProp,
-        default: () => NoElement
+        default: () => NoAnimation
       },
       /**
        * Animation to use for the overlay (backdrop) part of the modal. By default, it uses
@@ -121,7 +117,7 @@
        *
        * @param event - The click event.
        */
-      function emitOverlayClicked(event: MouseEvent) {
+      function emitOverlayClicked(event: Event) {
         emit('click:overlay', event);
       }
