+++ packages/x-components/src/components/panels/base-tabs-panel.vue
@@ -66,8 +66,8 @@
 
 <script lang="ts">
   import { defineComponent, ref } from 'vue';
-  import { NoElement } from '../no-element';
   import { AnimationProp } from '../../types';
+  import { NoAnimation } from '../animations';
 
   /**
    * Base Tabs Panel.
@@ -93,7 +93,7 @@
        */
       contentAnimation: {
         type: AnimationProp,
-        default: () => NoElement
+        default: () => NoAnimation
       },
       /**
        * The tab to be initially selected.
